+++ src/views/controls/nav-button.js
@@ -1,16 +1,33 @@
 /* eslint-env browser */
 import usage from './nav-button.usage.js';
 
+const props = `is not array? | {
+    name: undefined,
+    text: #.props has no 'content' ? is (string or number or boolean) ?: text,
+    content: undefined,
+    disabled: false,
+    href,
+    external,
+    onClick: undefined
+} | overrideProps()`;
+
 export default function(host) {
-    host.view.define('nav-button', function(el, config, data, context) {
-        const { name, content, disabled = false, onClick } = config;
-        const { text = '', href, external } = data || {};
+    host.view.define('nav-button', function(el, props, data, context) {
+        const {
+            name,
+            text,
+            content,
+            disabled,
+            href,
+            external,
+            onClick
+        } = props;
 
         if (name) {
             el.dataset.name = name;
         }
 
-        if (host.query(disabled, data, context)) {
+        if (disabled) {
             el.classList.add('disabled');
         } else if (typeof onClick === 'function') {
             el.addEventListener('click', (event) => onClick(el, data, context, event));
@@ -27,6 +44,7 @@ export default function(host) {
         }
     }, {
         tag: 'a',
+        props,
         usage
     });
 }