+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `@discoveryjs/discovery/utils` export
 - Renamed `Widget` class into `ViewModel`
+- Changed `nav-button` view to align with `button` view (`text`, `href` and `external` as config options, `disabled` is no longer treated as query)
 - Modified `PageRenderer#define()` and `ViewRenderer#define()` to allow specifying all options with a single config parameter:
   ```js
   // new signature
@@ -12,6 +13,7 @@
 - Removed `export default` in the core modules
 - Fixed `Model#pathToQuery()` method to correctly generate a query when part of the path contains a jora keyword
 - Fixed an edge case in the `table` view when rendering a single cell with an explicit column setup and row data containing non-object values
+- 
 
 ## 1.0.0-beta.88 (24-10-2024)